+++ source-files/js/_afterLoad.js
@@ -52,9 +52,16 @@ if (!(window.customElements && document.body.attachShadow)) {
   //
   document.getElementById('globVars').initializePlugin();
   document.getElementById('displayUtils').initializePlugin();
+  // document.getElementById('beepSounds').initializePlugin();
+  // document.getElementById('localCommandParser').initializePlugin();
+  // document.getElementById('remoteCommandParser').initializePlugin();
   document.getElementById('websocketPanel').initializePlugin();
   document.getElementById('navMenu').initializePlugin();
   document.getElementById('headerBar').initializePlugin();
+  // document.getElementById('serverForm').initializePlugin();
+  // document.getElementById('ircControlsPanel').initializePlugin();
+  // document.getElementById('ircServerPanel').initializePlugin();
+  document.getElementById('wallopsPanel').initializePlugin();
   document.getElementById('noticePanel').initializePlugin();
 
   /**
@@ -79,7 +86,12 @@ if (!(window.customElements && document.body.attachShadow)) {
   setInterval(() => {
     document.getElementById('errorPanel').timerTickHandler();
     document.getElementById('displayUtils').timerTickHandler();
+    // document.getElementById('beepSounds').timerTickHandler();
+    // document.getElementById('localCommandParser').timerTickHandler();
+    // document.getElementById('remoteCommandParser').timerTickHandler();
     document.getElementById('websocketPanel').timerTickHandler();
+    // document.getElementById('ircControlsPanel').timerTickHandler();
+    // document.getElementById('ircServerPanel').timerTickHandler();
     document.getElementById('manageChannelsPanel').timerTickHandler();
   }, 1000);