+++ source-files/web-components/irc-controls-panel.js
@@ -21,14 +21,12 @@
 // SOFTWARE.
 // ------------------------------------------------------------------------------
 //
-// This web component is a UI panel to display IRC Notice messages
-//
 // ------------------------------------------------------------------------------
 'use strict';
-window.customElements.define('notice-panel', class extends HTMLElement {
+window.customElements.define('irc-controls-panel', class extends HTMLElement {
   constructor () {
     super();
-    const template = document.getElementById('noticePanelTemplate');
+    const template = document.getElementById('ircControlsPanelTemplate');
     const templateContent = template.content;
     this.attachShadow({ mode: 'open' })
       .appendChild(templateContent.cloneNode(true));
@@ -36,25 +34,30 @@ window.customElements.define('notice-panel', class extends HTMLElement {
 
   showPanel = () => {
     this.shadowRoot.getElementById('panelVisibilityDivId').setAttribute('visible', '');
+    this.shadowRoot.getElementById('panelCollapsedDivId').setAttribute('visible', '');
   };
 
-  // this panel does not collapse, so close it.
   collapsePanel = () => {
-    this.shadowRoot.getElementById('panelVisibilityDivId').removeAttribute('visible');
+    this.shadowRoot.getElementById('panelVisibilityDivId').setAttribute('visible', '');
+    this.shadowRoot.getElementById('panelCollapsedDivId').removeAttribute('visible');
   };
 
   hidePanel = () => {
     this.shadowRoot.getElementById('panelVisibilityDivId').removeAttribute('visible');
+    this.shadowRoot.getElementById('panelCollapsedDivId').removeAttribute('visible');
   };
+
+  /**
+   * Called once per second as task scheduler, called from js/_afterLoad.js
+   */
+  // timerTickHandler = () => {
+  // };
 
   // ------------------
   // Main entry point
   // ------------------
-  initializePlugin = () => {
-    // console.log('notice-panel initializePlugin');
-    this.shadowRoot.getElementById('panelMessageDisplayId').value =
-      '12:01:26 MyNickname | This is an example NOTICE message\n';
-  };
+  // initializePlugin = () => {
+  // }; // initializePlugin()
 
   // add event listeners to connected callback
   // -------------------------------------------
@@ -63,43 +66,20 @@ window.customElements.define('notice-panel', class extends HTMLElement {
     // 1 of 2 Listeners on internal elements
     // -------------------------------------
 
-    // -------------------------------------
-    // Panel Close Button
-    // -------------------------------------
     this.shadowRoot.getElementById('closePanelButtonId').addEventListener('click', () => {
-      this.shadowRoot.getElementById('panelVisibilityDivId').removeAttribute('visible');
+      this.hidePanel();
     });
 
-    // -------------------------
-    // Erase button handler
-    // -------------------------
-    this.shadowRoot.getElementById('eraseButtonId').addEventListener('click', () => {
-      this.shadowRoot.getElementById('panelMessageDisplayId').value =
-        '\nTODO: fetch request to server for cache erase';
-    }); // panel erase button
-
-    // -------------------------
-    // Taller button handler
-    // -------------------------
-    this.shadowRoot.getElementById('tallerButtonId').addEventListener('click', () => {
-      const newRows =
-        parseInt(this.shadowRoot.getElementById('panelMessageDisplayId').getAttribute('rows')) + 5;
-      this.shadowRoot.getElementById('panelMessageDisplayId')
-        .setAttribute('rows', newRows.toString());
+    this.shadowRoot.getElementById('collapsePanelButtonId').addEventListener('click', () => {
+      if (this.shadowRoot.getElementById('panelCollapsedDivId').hasAttribute('visible')) {
+        this.collapsePanel();
+      } else {
+        this.showPanel();
+      }
     });
 
-    // -------------------------
-    // Normal button handler
-    // -------------------------
-    this.shadowRoot.getElementById('normalButtonId').addEventListener('click', () => {
-      this.shadowRoot.getElementById('panelMessageDisplayId').setAttribute('rows', '5');
-    });
-
-    // -------------------------------
-    // Clear message activity ICON by clicking on the main
-    // -------------------------------
-    this.shadowRoot.getElementById('panelDivId').addEventListener('click', function () {
-      // resetNotActivityIcon();
+    this.shadowRoot.getElementById('editServerButtonId').addEventListener('click', () => {
+      document.getElementById('serverForm').showPanel();
     });
 
     // -------------------------------------
@@ -128,19 +108,19 @@ window.customElements.define('notice-panel', class extends HTMLElement {
       }
     });
 
+    /**
+     * Global event listener on document object to implement changes to color theme
+     * @listens document:color-theme-changed
+     * @param {object} event.detail.theme - Color theme values 'light' or 'dark'
+     */
     document.addEventListener('color-theme-changed', (event) => {
       const panelDivEl = this.shadowRoot.getElementById('panelDivId');
-      const panelMessageDisplayEl = this.shadowRoot.getElementById('panelMessageDisplayId');
       if (event.detail.theme === 'light') {
-        panelDivEl.classList.remove('notice-panel-theme-dark');
-        panelDivEl.classList.add('notice-panel-theme-light');
-        panelMessageDisplayEl.classList.remove('global-text-theme-dark');
-        panelMessageDisplayEl.classList.add('global-text-theme-light');
+        panelDivEl.classList.remove('irc-controls-panel-theme-dark');
+        panelDivEl.classList.add('irc-controls-panel-theme-light');
       } else {
-        panelDivEl.classList.remove('notice-panel-theme-light');
-        panelDivEl.classList.add('notice-panel-theme-dark');
-        panelMessageDisplayEl.classList.remove('global-text-theme-light');
-        panelMessageDisplayEl.classList.add('global-text-theme-dark');
+        panelDivEl.classList.remove('irc-controls-panel-theme-light');
+        panelDivEl.classList.add('irc-controls-panel-theme-dark');
       }
     });
 
@@ -163,21 +143,6 @@ window.customElements.define('notice-panel', class extends HTMLElement {
         }
       } else {
         this.hidePanel();
-      }
-    });
-
-    /**
-     * Change size of textArea elements to fit page
-     * @listens document:resize-custom-elements
-     */
-    document.addEventListener('resize-custom-elements', () => {
-      if (window.globals.webState.dynamic.inputAreaCharWidthPx) {
-        const calcInputAreaColSize = document.getElementById('displayUtils').calcInputAreaColSize;
-        // pixel width mar1 is reserved space on edges of input area at full screen width
-        const mar1 = window.globals.webState.dynamic.commonMargin;
-        // set width of input area elements
-        this.shadowRoot.getElementById('panelMessageDisplayId')
-          .setAttribute('cols', calcInputAreaColSize(mar1));
       }
     });