+++ source-files/web-components/logout-panel.js
@@ -1,0 +1,162 @@
+// MIT License
+//
+// Copyright (c) 2021 Dave Bolenbaugh
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+// --------------------------------------------------------------------------------
+//   Logout Panel, notify user that IRC is still connected and will
+//   remain connected after web page user logout.
+// --------------------------------------------------------------------------------
+'use strict';
+window.customElements.define('logout-panel', class extends HTMLElement {
+  constructor () {
+    super();
+    const template = document.getElementById('logoutPanelTemplate');
+    const templateContent = template.content;
+    this.attachShadow({ mode: 'open' })
+      .appendChild(templateContent.cloneNode(true));
+  }
+
+  /**
+   * Make panel visible
+   */
+  showPanel = () => {
+    this.shadowRoot.getElementById('panelVisibilityDivId').setAttribute('visible', '');
+  };
+
+  /**
+   * This panel does not collapse, instead close it.
+   */
+  collapsePanel = () => {
+    this.shadowRoot.getElementById('panelVisibilityDivId').removeAttribute('visible');
+  };
+
+  /**
+   * Hide panel
+   */
+  hidePanel = () => {
+    this.shadowRoot.getElementById('panelVisibilityDivId').removeAttribute('visible');
+  };
+
+  handleLogoutRequest = () => {
+    if (((window.globals.ircState.ircConnected) && (window.globals.webState.webConnected)) ||
+      (!window.globals.webState.webConnected)) {
+      this.showPanel();
+      document.dispatchEvent(new CustomEvent('global-scroll-to-top'));
+    } else {
+      // removed user data stored locally.
+      window.localStorage.clear();
+      window.location.href = '/logout';
+    }
+  };
+
+  // initializePlugin () {
+  // }
+
+  connectedCallback () {
+    // -------------------------------------
+    // 1 of 2 Listeners on internal elements
+    // -------------------------------------
+
+    this.shadowRoot.getElementById('closePanelButtonId').addEventListener('click', () => {
+      this.hidePanel();
+    });
+
+    this.shadowRoot.getElementById('cancelButtonId').addEventListener('click', () => {
+      this.hidePanel();
+    });
+
+    this.shadowRoot.getElementById('webLogoutButtonId').addEventListener('click', () => {
+      // removed user data stored locally.
+      window.localStorage.clear();
+      window.location.href = '/logout';
+    });
+
+    // -------------------------------------
+    // 2 of 2 Listeners on global events
+    // -------------------------------------
+
+    /**
+     * Event to collapse all panels. This panel does not collapse so it is hidden
+     * @listens document:collapse-all-panels
+     * @property {string|string[]} event.detail.except - No action if listed as exception
+    */
+    document.addEventListener('collapse-all-panels', (event) => {
+      if ((event.detail) && (event.detail.except)) {
+        if (typeof event.detail.except === 'string') {
+          // this.id assigned in html/_index.html
+          if (event.detail.except !== this.id) {
+            this.collapsePanel();
+          }
+        } else if (Array.isArray(event.detail.except)) {
+          if (event.detail.except.indexOf(this.id) < 0) {
+            this.collapsePanel();
+          }
+        }
+      } else {
+        this.hidePanel();
+      }
+    });
+
+    /**
+     * Hide panel (not visible) unless listed as exception.
+     * @listens document:hide-all-panels
+     * @property {string|string[]} event.detail.except - No action if listed as exception
+     */
+    document.addEventListener('hide-all-panels', (event) => {
+      if ((event.detail) && (event.detail.except)) {
+        if (typeof event.detail.except === 'string') {
+          // this.id assigned in html/_index.html
+          if (event.detail.except !== this.id) {
+            this.hidePanel();
+          }
+        } else if (Array.isArray(event.detail.except)) {
+          if (event.detail.except.indexOf(this.id) < 0) {
+            this.hidePanel();
+          }
+        }
+      } else {
+        this.hidePanel();
+      }
+    });
+
+    /**
+     * Make panel visible unless listed as exception.
+     * Special case, only open panel if debug === true
+     * @listens document:show-all-panels
+     * @param {string|string[]} event.detail.except - No action if listed as exception
+     */
+    document.addEventListener('show-all-panels', (event) => {
+      if ((event.detail) && (event.detail.except)) {
+        if (typeof event.detail.except === 'string') {
+          // this.id assigned in html/_index.html
+          if ((event.detail.except !== this.id) && (event.detail.debug)) {
+            this.showPanel();
+          }
+        } else if (Array.isArray(event.detail.except)) {
+          if ((event.detail.except.indexOf(this.id) < 0) && (event.detail.debug)) {
+            this.showPanel();
+          }
+        }
+      } else {
+        if (event.detail.debug) this.showPanel();
+      }
+    });
+  } // connectedCallback()
+});